glasses_backend/requirements.txt
deepface
numpy
retina-face
opencv-python
google-generativeai
python-dotenv
google-cloud-vision
fastapi[standard]
python-multipart
uvicorn
pillow
flask
google-genai
tensorflow
keras
python-dateutil